Ignore WPS202, WPS402 and WPS210
WPS202 limits the amount of module members which does not necessarily
make sense to all cases. WPS402 limits the amount of noqa comments which
can be hard to avoid sometimes. WPS210 limits the amount of local
variables in a method which, in my opinion, is a bad measure of
complexity.